feat: initial implementation of slect
Implements the following loose specification for an extension of Clap. We'll work on this progressively.

So, I think this approach makes the most sense.

We will define a (derive) macro that allows us to write...

pub struct MySubcommand {
    #[select(SubcommandA, SubcommandB, SubcommandC)]
      pub extra_args: Vec<String>
}

This is going to then going to do the following:

1. It's going to create an additional field --help-all which if present will always call the normal Clap help for the struct, but also call all of helps for the possibilities in Select. This will all be wrapped up in a new implementation for a function help_all on the MySubcommand struct.
2. It's also going to implement a method select which parses all of the selections into an Option<Selections> tuple. (Here we will also add prefixing logic)>

I suppose for prefixing, this form would be nice.

 pub extra_args: Vec<String>